cleaned up FeefForward stuff
seperated classes plus made it shorter and neater
